server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "simple music app API."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.0",
    "fastapi[standard]>=0.112.2",
    "sqlalchemy",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.9.0",
    "pydantic-settings>=2.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
